requirement.txt
streamlit>=1.24.0
tensorflow-cpu==2.15.0
numpy==1.24.3
pandas>=1.5.3
scikit-learn>=1.2.2
tensorboard==2.15.1
matplotlib>=3.7.1

